+++ goldmane/pkg/internal/utils/logging.go
@@ -21,6 +21,9 @@ import (
 	"github.com/projectcalico/calico/libcalico-go/lib/logutils"
 )
 
+// ConfigureLogging configures the logging framework. The logging level that will
+// be used is passed in as a parameter. Otherwise, it will default to WARN
+// The output will be set to STDOUT and the format is TextFormat
 func ConfigureLogging(logLevel string) {
 	// Install a hook that adds file/line number information.
 	logutils.ConfigureFormatter("goldmane")
